Move embed view to separate file
